split robot and world into classes
