+++ client/src/global-types.ts
@@ -1,0 +1,5 @@
+export enum Terrian {P = 'Plains', M = 'Mountains', C = 'Crevasse'};
+export type Command = 'L' | 'R' | 'F' | 'B';
+export type Direction = 'N' | 'S' | 'E' | 'W';
+export type Location = [number,number];
+export type World = Array<Array<Terrian>>;
